+++ __registry__/index.tsx
@@ -11,12 +11,12 @@ export const Index: Record<string, any> = {
       type: "registry:ui",
       registryDependencies: ["command","skeleton"],
       files: [{
-        path: "registry/default/akasha-ui/auto-complete.tsx",
+        path: "registry/default/akasha-ui/autocomplete.tsx",
         type: "registry:ui",
         target: ""
       }],
       categories: undefined,
-      component: React.lazy(() => import("@/registry/default/akasha-ui/auto-complete.tsx")),
+      component: React.lazy(() => import("@/registry/default/akasha-ui/autocomplete.tsx")),
       source: "",
       meta: undefined,
     },
@@ -189,14 +189,14 @@ export const Index: Record<string, any> = {
       name: "autocomplete-demo",
       description: "",
       type: "registry:example",
-      registryDependencies: ["auto-complete"],
+      registryDependencies: ["autocomplete"],
       files: [{
-        path: "registry/default/examples/auto-complete-demo.tsx",
+        path: "registry/default/examples/autocomplete-demo.tsx",
         type: "registry:example",
         target: ""
       }],
       categories: undefined,
-      component: React.lazy(() => import("@/registry/default/examples/auto-complete-demo.tsx")),
+      component: React.lazy(() => import("@/registry/default/examples/autocomplete-demo.tsx")),
       source: "",
       meta: undefined,
     },
@@ -204,14 +204,14 @@ export const Index: Record<string, any> = {
       name: "autocomplete-multiple-demo",
       description: "",
       type: "registry:example",
-      registryDependencies: ["auto-complete"],
+      registryDependencies: ["autocomplete"],
       files: [{
-        path: "registry/default/examples/auto-complete-multiple-demo.tsx",
+        path: "registry/default/examples/autocomplete-multiple-demo.tsx",
         type: "registry:example",
         target: ""
       }],
       categories: undefined,
-      component: React.lazy(() => import("@/registry/default/examples/auto-complete-multiple-demo.tsx")),
+      component: React.lazy(() => import("@/registry/default/examples/autocomplete-multiple-demo.tsx")),
       source: "",
       meta: undefined,
     },
