+++ content/docs/components/autocomplete.mdx
@@ -5,7 +5,7 @@ component: true
 ---
 
 <ComponentPreview
-  name="auto-complete-demo"
+  name="autocomplete-demo"
   description="Autocomplete with selection of single option example"
 />
 
@@ -21,7 +21,7 @@ component: true
 <TabsContent value="cli">
 
 ```bash
-npx shadcn@latest add "https://akasha-ui.pages.dev/r/styles/default/auto-complete.json"
+npx shadcn@latest add "https://akasha-ui.pages.dev/r/styles/default/autocomplete.json"
 ```
 
 </TabsContent>
@@ -38,7 +38,7 @@ npm install next
 
 <Step>Copy and paste the following code into your project.</Step>
 
-<ComponentSource name="auto-complete" />
+<ComponentSource name="autocomplete" />
 
 <Step>Update the import paths to match your project setup.</Step>
 
@@ -51,7 +51,7 @@ npm install next
 ## Usage
 
 ```tsx
-import { Autocomplete } from "<path-to-ui-components>/auto-complete";
+import { Autocomplete } from "<path-to-ui-components>/autocomplete";
 ```
 
 ```tsx
@@ -70,4 +70,4 @@ import { Autocomplete } from "<path-to-ui-components>/auto-complete";
 
 Passing `multiple={true}` will allow users to select multiple options.
 
-<ComponentPreview name="auto-complete-multiple-demo" />
+<ComponentPreview name="autocomplete-multiple-demo" />
